new snippet formatting code
This code handles a richer form of MultiSpan that includes
labels.

Unlike the old code, this code doesn't try to emit things as it goes,
but rather builds things up into a buffer and gradually adjusts it.  It
also produces "styled" output that allows to use a richer color palette,
e.g. highlighting the `~` distinctly from other things.