add labels to multispan, permit overlapping spans
The current MultiSpan had a bunch of logic for removing overlapping
spans, grouping spans, and so forth. The newer MultiSpan just permits
overlapping spans, and allows each span to have a label -- this is
partly because, once you add labels, it is not obvious how to group
spans together anymore, and so forth. The new snippet rendering code has
similarly been extended to accommodate spans that overlap.