extend DiagnosticBuilder with `label_span`
When you create the diagnostic buidler, you supply the
primary span (that is the one that is given as the
"main" error point). If you do nothing else, that span will
be highlighted in your snippet, just like in the old days.

But if you prefer, you can call `label_span` to manually add
the spans you want labeled, along with the labels you want.

I had `label_span` take by-value and return by-value. I suspect the
other methods should also be converted in this fashion; it seemed to
make for cleaner code. But maybe not.