cleanup the logic around divergence and match
The old code was (incorrectly) using a **diverging type var** to start,
even when the match had >0 arms. This would lead to incorrect fallback.
Also, the way the warnings are setup, we would not give an "unreachable
expression" warning for something like `match (return) { }`.

(I think this problem may be a bit more general, actually; there are
probably missing calls for other places, but I'm not going to worry
about it right now. Also, I think that the handling of the diverges
flags for patterns is wrong: patterns are always extracting data from an
existing lvalue, so they should never set the diverging flag, but they
do (but then it gets masked, so maybe it all works out).)