incorporate param-env into caching
This is a naive integration -- it will lead to less effective caching
than it could because we don't adjust the universe index. However, it
also fixes a pre-existing bug -- in particular, we could wind up
storing a cache entry that is specific to a param env in the local
cache; evaluation then clears the param env, but still hits the local
cache. This defeats the purpose of having evaluation clear the param
env.