replace custom visitor with invoking the outlives function
In particular, this is the justification, essentially, for ignoring
the non-upvar closure parameters -- outlives also ignores them, which
means that the rest of the compiler can't tell the difference. The
only way that they get observed is via auto traits that could perhaps
use an outlives obligation, or via trans when they are erased.